Kubernetes_part/kubernetes/pub_kubernetes/requirements.txt
Django
redis
requests
python-dotenv
psycopg2-binary
django-cors-headers